+++ CHANGES.md
@@ -6,10 +6,15 @@ Version 0.18.3
 
 To be released.
 
+ -  Fixed `Swarm<T>`'s bug that it had thrown `PingTimeoutException` if any
+    peer in configured `SwarmOptions.StaticPeers` is unreachable.
+    [[#1550], [#1551], [#1562]]
  -  Fixed a bug where `Swarm<T>` did not removed failed block demands from the
     `BlockDemandTable`.  [[#1549], [#1562]]
 
 [#1549]: https://github.com/planetarium/libplanet/pull/1549
+[#1550]: https://github.com/planetarium/libplanet/issues/1550
+[#1551]: https://github.com/planetarium/libplanet/pull/1551
 [#1562]: https://github.com/planetarium/libplanet/pull/1561